Split validation keys
