+++ libmamba/CMakeLists.txt
@@ -180,6 +180,7 @@ set(
     ${LIBMAMBA_SOURCE_DIR}/specs/version_spec.cpp
     # Artifacts validation
     ${LIBMAMBA_SOURCE_DIR}/validation/errors.cpp
+    ${LIBMAMBA_SOURCE_DIR}/validation/keys.cpp
     # C++ wrapping of libcurl
     ${LIBMAMBA_SOURCE_DIR}/core/curl.cpp
     # C++ wrapping of compression libs (zstd and bzlib)
@@ -287,6 +288,7 @@ set(
     ${LIBMAMBA_INCLUDE_DIR}/mamba/specs/version_spec.hpp
     # Artifacts validation
     ${LIBMAMBA_INCLUDE_DIR}/mamba/validation/errors.hpp
+    ${LIBMAMBA_INCLUDE_DIR}/mamba/validation/keys.hpp
     # Core API (low-level)
     ${LIBMAMBA_INCLUDE_DIR}/mamba/core/activation.hpp
     ${LIBMAMBA_INCLUDE_DIR}/mamba/core/channel_context.hpp
