+++ libmamba/include/mamba/core/validate.hpp
@@ -16,6 +16,7 @@
 #include "mamba/core/timeref.hpp"
 #include "mamba/core/util.hpp"
 #include "mamba/fs/filesystem.hpp"
+#include "mamba/validation/keys.hpp"
 
 namespace mamba::validation
 {
@@ -79,93 +80,6 @@ namespace mamba::validation
     );
 
     void check_timestamp_metadata_format(const std::string& ts);
-
-    /**
-     * Representation of the public part of a cryptographic key pair.
-     */
-    struct Key
-    {
-        std::string keytype = "";
-        std::string scheme = "";
-        std::string keyval = "";
-
-        static Key from_ed25519(std::string keyval)
-        {
-            return { "ed25519", "ed25519", std::move(keyval) };
-        }
-    };
-
-    void to_json(nlohmann::json& j, const Key& k);
-    void from_json(const nlohmann::json& j, Key& k);
-
-
-    /**
-     * Representation of a role signature.
-     *
-     * Optional 'pgp_trailer' will trigger special handling during verification to conform to
-     * OpenPGP RFC4880.
-     */
-    struct RoleSignature
-    {
-        std::string keyid = "";
-        std::string sig = "";
-        std::string pgp_trailer = "";
-    };
-
-    void to_json(nlohmann::json& j, const RoleSignature& rs);
-    void from_json(const nlohmann::json& j, RoleSignature& rs);
-
-    bool operator<(const RoleSignature& rs1, const RoleSignature& rs2);
-
-
-    /**
-     * Store key IDs and threshold for a role.
-     *
-     * Key ID can be a hash of Key, or just its public key value.
-     */
-    struct RoleKeys
-    {
-        std::vector<std::string> keyids;
-        std::size_t threshold;
-    };
-
-    void to_json(nlohmann::json& j, const RoleKeys& rk);
-    void from_json(const nlohmann::json& j, RoleKeys& rk);
-
-
-    /**
-     * Store key values and threshold for role. Assumes key scheme/type is `ed25519`.
-     */
-    struct RolePubKeys
-    {
-        std::vector<std::string> pubkeys;
-        std::size_t threshold;
-
-        RoleKeys to_role_keys() const;
-    };
-
-    void to_json(nlohmann::json& j, const RolePubKeys& rk);
-    void from_json(const nlohmann::json& j, RolePubKeys& rk);
-
-
-    /**
-     * Store full keys and threshold for role.
-     */
-    struct RoleFullKeys
-    {
-        RoleFullKeys() = default;
-        RoleFullKeys(const std::map<std::string, Key>& keys_, const std::size_t& threshold_);
-
-        std::map<std::string, Key> keys;
-        std::size_t threshold;
-
-        std::map<std::string, Key> to_keys() const;
-        RoleKeys to_roles() const;
-    };
-
-    void to_json(nlohmann::json& j, const RoleFullKeys& r);
-    void from_json(const nlohmann::json& j, RoleFullKeys& r);
-
 
     /**
      * Base class for spec implementations.
